pyproject.toml
[project]
name = "prana-chain-openenv"
version = "0.1.0"
description = "OpenEnv environment for emergency oxygen triage logistics."
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.2",
  "pydantic>=2.0.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.24.0",
  "openai>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["prana_chain", "prana_chain.server"]

[project.scripts]
server = "server.app:main"
